Working sqw_eval in Python
Add Matlab outputs to call_python.cpp (assume only one array is outputted).
Add pythonFunctionWrapper class in Matlab.
Changed DataTypes and FunctionWrapper to return this Matlab object to be
  passed to the Matlab sqw_eval etc.
When it goes out of focus in Matlab, its destructor calls the mex
  pythonRemoveFuncKey.cpp to remove the reference to the function in
  the module-global dict.
Add an example script to call sqw_eval() with a Python function


Former-commit-id: f14847c5d101ff5e848d6036eb0d46a3199ae0fa